Move `db.*` analyzers to `sqlalchemy.*`
The database analyzers are more specifically named now to disambiguate between
generic SQLAlchemy analyzers and more speciifc ones like `snowflake.columns` or
`bigquery.access`.

This approach will allow different analyzers to coexist with eachother, and
allow clients to pick which they wish to pay attention to.

Closes #112
